basic structure created
